modified:   actividades.html 	modified:   blog.html 	modified:   compilar_web.sh 	modified:   index.html 	modified:   integrantes.html 	modified:   investigacion.html 	modified:   publicaciones.php 	modified:   resultados.html 	modified:   seminario/2005.html 	modified:   seminario/2006.html 	modified:   seminario/2007.html 	modified:   seminario/2008.html 	modified:   seminario/2009.html 	modified:   seminario/2010.html 	modified:   src/header.html

+++ compilar_web.sh
@@ -19,9 +19,24 @@ do
 
 done
 
-rm *~
-rm src/*~
-rm seminario/*~
-rm src/seminario/*~
+if [ -f "*~" ];
+then 
+	rm *~ 
+fi
+
+if [ -f "src/*~" ];
+then 
+	rm src/*~ 
+fi
+
+if [ -f "seminario/*~" ];
+then 
+	rm seminario/*~ 
+fi
+
+if [ -f "src/seminario/*~" ];
+then
+	rm src/seminario/*~ 
+fi
 
 echo "Finalizado"
